Modified inputs to freq tuned foam emissivity model
Removed additional freq input (band variable) because it could be set independently of the frequency used in the rest of the model and it was limited to 6 discrete freq. Now the routne uses the same input freq as the rest of the model, and it interpolates the foam model parameters at this freq from the 6 discrete freq.