Optimized multilayer foam emissivity models
Removed computation of dielectric constant model for sea water from inside the foam emission routine. Now the dielectric constant for sea water is computed in the main programonly  when SSS and SST. Before, the dielectric constant was recompted every time the foam emission is called, which can be millions of times when integrating on the large scale slopes, and only the incidence angle changes (hence no need to recompute the dielectric constant). Also started to add header to foam routines (TBC).